Improve rewards transaction details
