+++ components/Map/Map.js
@@ -114,7 +114,6 @@ const CoverageMap = () => {
 
       setSelectedTxnHotspot(targetHotspot)
       setSelectedTxnParticipants(witnesses)
-      return
     } else if (
       selectedTxn?.type === 'assert_location_v1' ||
       selectedTxn?.type === 'assert_location_v2'
@@ -123,16 +122,13 @@ const CoverageMap = () => {
       const targetHotspot = await fetchHotspot(target)
       setSelectedTxnHotspot(targetHotspot)
       setSelectedTxnParticipants([])
-      return
     } else if (selectedTxn?.type === 'consensus_group_v1') {
       const members = await fetchConsensusHotspots(txn.height)
       setSelectedTxnHotspot(undefined)
       setSelectedTxnParticipants(members)
-      return
     } else {
       setSelectedTxnHotspot(undefined)
       setSelectedTxnParticipants([])
-      return
     }
   }, [selectedTxn])
 
@@ -194,11 +190,6 @@ const CoverageMap = () => {
       <HotspotDetailLayer
         hotspot={selectedHotspot || selectedTxnHotspot}
         witnesses={selectedHotspot?.witnesses || selectedTxnParticipants || []}
-        members={
-          selectedTxn?.type === 'consensus_group_v1'
-            ? selectedTxnParticipants
-            : []
-        }
       />
       <ValidatorsLayer
         validators={validators}
